requirements.txt
gspread
oauth2client
python-quickbooks
numpy==2.2.5
pandas==2.2.3
pyarrow==20.0.0
streamlit==1.45.1
pillow==11.2.1
protobuf==6.30.2
requests==2.32.3
click==8.2.1
gitpython==3.1.44
altair==5.5.0
pydeck==0.9.1
cachetools==5.5.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5
typing-extensions==4.13.2
pywin32